production

+++ routes/categories.js
@@ -3,6 +3,7 @@ const router = express.Router()
 const { validatorName, Category } =require('../models/category')
 const auth = require('../middleware/auth')
 const admin = require('../middleware/admin')
+const mongoose = require("mongoose");
 
 
 
@@ -26,17 +27,22 @@ router.get('/', async (req, res) => {
 
 })
 
-router.get('/:id',auth, async (req, res) => {
-
-    const category = await Category.findById(req.params.id)
+router.get('/:id', async (req, res) => {
+    if (mongoose.Types.ObjectId.isValid(req.params.id) === false) {
+        return res.status(404).send('Bunday id mavjud emas')
+    }
+    let category = await Category.findById(req.params.id)
     if (!category) {
-        return res.status(400).send('Bunday id li category mavjud emas')
+        return res.status(404).send('Bunday id li category mavjud emas')
     }
     res.send(category)
 
 })
 
 router.put('/:id',auth, async (req, res) => {
+    if (mongoose.Types.ObjectId.isValid(req.params.id) === false) {
+        return res.status(404).send('Bunday id mavjud emas')
+    }
     const { error } = validatorName(req.body)
     if (error) {
         return res.status(400).send(error.details[0].message)
@@ -45,15 +51,18 @@ router.put('/:id',auth, async (req, res) => {
         name: req.body.name
     }, { new: true })
     if (!category) {
-        return res.status(400).send('Bunday id li category mavjud emas')
+        return res.status(404).send('Bunday id li category mavjud emas')
     }
     res.send(category)
 })
 
 router.delete('/:id',[ auth, admin ], async (req, res) => {
+    if (mongoose.Types.ObjectId.isValid(req.params.id) === false) {
+        return res.status(404).send('Bunday id mavjud emas')
+    }
     const category = await Category.findByIdAndDelete(req.params.id)
     if (!category) {
-        return res.status(400).send('Bunday id li category mavjud emas')
+        return res.status(404).send('Bunday id li category mavjud emas')
     }
     res.send(category)
 })
